Release 3.9 1.18 support Added /damage Added /jesus Added server-only mode (beta feature) Added permissions for boolean gamerules. Fixed incompatibility with Fabric 0.12.x Fixed big items looking whack when Fix Item Seams client tweak was enabled. Fixed mod incompatibility causing chairs to crash clients and servers. Fixed fake players not showing skin layers. Fixed certain enchantment levels crashing the game when hovering over items that have them. Fixed hopper minecarts taking items from inventories crashing the game. Fixed weird behaviour regarding reach. Fixed player data getting reset upon death.
Many more fixes that I just forgot to mention

+++ src/main/java/com/ptsmods/morecommands/commands/server/elevated/EnchantMaxCommand.java
@@ -4,7 +4,6 @@ import com.mojang.brigadier.CommandDispatcher;
 import com.ptsmods.morecommands.miscellaneous.Command;
 import net.minecraft.item.ItemStack;
 import net.minecraft.server.command.ServerCommandSource;
-import net.minecraft.util.Formatting;
 import net.minecraft.util.registry.Registry;
 
 public class EnchantMaxCommand extends Command {
@@ -15,7 +14,7 @@ public class EnchantMaxCommand extends Command {
             if (stack.isEmpty()) sendError(ctx, "You're not holding an item.");
             else {
                 stack.getEnchantments().clear();
-                Registry.ENCHANTMENT.stream().filter(enchantment -> !enchantment.isCursed()).forEach(enchantment -> stack.addEnchantment(enchantment, Integer.MAX_VALUE));
+                Registry.ENCHANTMENT.stream().filter(enchantment -> !enchantment.isCursed()).forEach(enchantment -> stack.addEnchantment(enchantment, Short.MAX_VALUE));
                 sendMsg(ctx, "Your item has been enchanted.");
                 return stack.getEnchantments().size();
             }
